Separating more of the data fetching from the UI to have a proper structure
